refactor: add .query, .input and .type suffix
