+++ src/routes/graphql/queries/member.query.ts
@@ -1,5 +1,5 @@
 import { GraphQLNonNull } from "graphql";
-import { MemberType, MemberTypeIdEnum } from "../types/member.js";
+import { MemberType, MemberTypeIdEnum } from "../types/member.type.js";
 
 export const MemberQuery = {
   type: MemberType,
