+++ src/routes/graphql/queries/members.query.ts
@@ -1,5 +1,5 @@
 import { GraphQLList, GraphQLNonNull } from "graphql";
-import { MemberType } from "../types/member.js";
+import { MemberType } from "../types/member.type.js";
 
 export const MembersQuery = {
   type: new GraphQLList(new GraphQLNonNull(MemberType)),
