+++ src/routes/graphql/schemas.ts
@@ -1,7 +1,7 @@
 import { Type } from '@fastify/type-provider-typebox';
 import { GraphQLSchema } from 'graphql';
-import { RootQueryType } from './queries/root-query.js';
-import { Mutations } from './mutations.js';
+import { RootQueryType } from './queries/root.query.js';
+import { Mutations } from './mutations/mutations.js';
 
 export const gqlResponseSchema = Type.Partial(
   Type.Object({
