+++ src/routes/graphql/types/member.type.ts
