+++ src/routes/graphql/types/uuid.type.ts
